CB-API/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
black==23.10.1
cffi==1.16.0
click==8.1.7
cryptography==39.0.2
ecdsa==0.18.0
fastapi==0.103.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.11.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyMySQL==1.0.3
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.20.0
